[PRTL-2994] 10k - Filter Row - move action buttons (#2815)
